split out history schema, and remove `cellId`, which is not used anywhere
