+++ src/eval-frame/components/panes/__tests__/console-pane.test.js
@@ -26,7 +26,6 @@ describe("ConsolePaneUnconnected React component", () => {
     props = {
       history: [
         {
-          cellId: 0,
           lastRan: 1533078293981,
           content: "var a = 3"
         }
@@ -58,12 +57,10 @@ describe("ConsolePaneUnconnected React component", () => {
   it("always renders correct number of HistoryItem inside history-items", () => {
     props.history = [
       {
-        cellId: 0,
         lastRan: 1533078293981,
         content: "var a = 3"
       },
       {
-        cellId: 1,
         lastRan: 1533078293981,
         content: "var b = 3"
       }
@@ -84,7 +81,6 @@ describe("HistoryPane mapStateToProps", () => {
     state = {
       history: [
         {
-          cellId: 0,
           lastRan: 1533078293981,
           content: "var a = 3"
         }
